SE R2 Astromech. Moved R2-D2's ability to R2 Astromech class since they share the same ability.
